Move versioning exceptions into semantic_version class
