+++ dbt_platform_helper/providers/semantic_version.py
@@ -1,8 +1,21 @@
 import re
 from typing import Union
 
-from dbt_platform_helper.providers.validation import IncompatibleMajorVersionException
-from dbt_platform_helper.providers.validation import IncompatibleMinorVersionException
+from dbt_platform_helper.providers.validation import ValidationException
+
+
+class IncompatibleMajorVersionException(ValidationException):
+    def __init__(self, app_version: str, check_version: str):
+        super().__init__()
+        self.app_version = app_version
+        self.check_version = check_version
+
+
+class IncompatibleMinorVersionException(ValidationException):
+    def __init__(self, app_version: str, check_version: str):
+        super().__init__()
+        self.app_version = app_version
+        self.check_version = check_version
 
 
 class SemanticVersion:
